refactor qureyCTRL tests
