+++ src/specs/gnocchi-datasource-specs.ts
@@ -329,5 +329,131 @@ describe('GnocchiDatasource', function() {
       expect(results.length).to.be(0);
     });
   });
+
+
+  // QueryCtrl tests
+  var resources = [
+    {
+      "created_by_project_id": "8a722a26-e0a0-4993-b283-76925b7b02de",
+      "created_by_user_id": "5587ebf3-58a5-42eb-8024-ef756e09a552",
+      "ended_at": null,
+      "id": "cba8d3d5-d5e1-4692-bcfe-d77feaf01d7e",
+      "metrics": {
+        "temperature": "86adbe6c-22d7-4a86-9ab7-e8d112f6cb79",
+        "cpu_util": "ccdd3d2c-7f83-42a0-9280-49e0791349dd"
+      },
+      "project_id": "bd3a1e52-1c62-44cb-bf04-660bd88cd74d",
+      "revision_end": null,
+      "revision_start": "2015-09-10T08:00:25.690667+00:00",
+      "started_at": "2015-09-10T08:00:25.690654+00:00",
+      "type": "generic",
+      "user_id": "bd3a1e52-1c62-44cb-bf04-660bd88cd74d"
+    },
+    {
+      "created_by_project_id": "8a722a26-e0a0-4993-b283-76925b7b02de",
+      "created_by_user_id": "5587ebf3-58a5-42eb-8024-ef756e09a552",
+      "ended_at": null,
+      "id": "9b4f6da9-4e67-4cfd-83bd-fb4b8bcc8dd8",
+      "metrics": {},
+      "project_id": "bd3a1e52-1c62-44cb-bf04-660bd88cd74d",
+      "revision_end": null,
+      "revision_start": "2015-09-10T08:00:25.690667+00:00",
+      "started_at": "2015-09-10T08:00:25.690654+00:00",
+      "type": "generic",
+      "user_id": "bd3a1e52-1c62-44cb-bf04-660bd88cd74d"
+    },
+  ];
+
+  var metrics = [
+    {
+      "id": "b8c73d22-d944-47d9-9d84-1e7f618c25e1",
+      "archive_policy": {},  // shouldn't be empty be too big and useless for test
+      "created_by_project_id": "8a722a26-e0a0-4993-b283-76925b7b02de",
+      "created_by_user_id": "5587ebf3-58a5-42eb-8024-ef756e09a552",
+      "name": "temperature",
+      "resource_id": "cba8d3d5-d5e1-4692-bcfe-d77feaf01d7e"
+    },
+    {
+      "id": "86adbe6c-22d7-4a86-9ab7-e8d112f6cb79",
+      "archive_policy": {},  // shouldn't be empty be too big and useless for test
+      "created_by_project_id": "8a722a26-e0a0-4993-b283-76925b7b02de",
+      "created_by_user_id": "5587ebf3-58a5-42eb-8024-ef756e09a552",
+      "name": "cpu_util",
+      "resource_id": "cba8d3d5-d5e1-4692-bcfe-d77feaf01d7e"
+    }
+  ];
+
+
+  describe('suggestMetricIDs', function() {
+    var results;
+    it('should return metric ids', function() {
+      $httpBackend.expect('GET', "/v1/metric").respond(metrics);
+      ds.performSuggestQuery("foobar", "metrics", {}).then(function(data) { results = data ; });
+      $httpBackend.flush();
+      expect(results.length).to.be(2);
+      expect(results[0]).to.be('b8c73d22-d944-47d9-9d84-1e7f618c25e1');
+      expect(results[1]).to.be('86adbe6c-22d7-4a86-9ab7-e8d112f6cb79');
+    });
+  });
+
+  describe('suggestResourceIDs', function() {
+    var results;
+    it('should return resource ids', function() {
+      $httpBackend.expect('GET', "/v1/resource/generic").respond(resources);
+      ds.performSuggestQuery("foobar", "resources", {}).then(function(data) { results = data ; });
+      $httpBackend.flush();
+      expect(results.length).to.be(2);
+      expect(results[0]).to.be('cba8d3d5-d5e1-4692-bcfe-d77feaf01d7e');
+      expect(results[1]).to.be('9b4f6da9-4e67-4cfd-83bd-fb4b8bcc8dd8');
+    });
+  });
+
+  describe('suggestMetricNames', function() {
+    var results;
+    it('should return resource ids', function() {
+      var target = {'resource_id': 'cba8d3d5-d5e1-4692-bcfe-d77feaf01d7e', 'queryMode': 'resource'};
+      $httpBackend.expect('GET', "/v1/resource/generic/cba8d3d5-d5e1-4692-bcfe-d77feaf01d7e").respond(resources[0]);
+      ds.performSuggestQuery("foobar", "metric_names", target).then(function(data) { results = data ; });
+      $httpBackend.flush();
+      expect(results.length).to.be(2);
+      expect(results[0]).to.be('temperature');
+      expect(results[1]).to.be('cpu_util');
+    });
+  });
+
+  describe('validate query success', function() {
+    it('no target error', function() {
+      var target = {'resource_query': '{"=": {"id": "foobar"}}',
+                    'queryMode': 'resource_search', 'metric_name': 'cpu_util'};
+      $httpBackend.expect('POST', "/v1/search/resource/generic").respond([]);
+      $httpBackend.expect('POST', "/v1/search/resource/generic").respond([]);
+      var error = ds.validateTarget(target, false);
+      $httpBackend.flush();
+      expect(error).to.be(undefined);
+    });
+  });
+
+  describe('validate query missing field', function() {
+    it('resource', function() {
+      var target = {'resource_id': '', 'queryMode': 'resource', 'metric_name': ''};
+      var error = ds.validateTarget(target, false);
+      expect(error).to.be("Missing or invalid fields: Resource ID, Metric name");
+    });
+
+    it('metric', function() {
+      var target = {'metric_id': '', 'queryMode': 'metric'};
+      var error = ds.validateTarget(target, false);
+      expect(error).to.be("Missing or invalid fields: Metric ID");
+    });
+
+    it('resource_search', function() {
+      var target = {'resource_search': '', 'queryMode': 'resource_search', 'metric_name': ''};
+      var error = ds.validateTarget(target, false);
+      expect(error).to.be("Missing or invalid fields: Query, Metric name");
+    });
+  });
+
+
+
 });
 ;
